Feat: 식재료 입력, 식재료 보기 api 구현

+++ src/main/java/com/team200/graduation_project/domain/recipe/entity/RecipeStep.java
@@ -17,7 +17,7 @@ public class RecipeStep {
     @JoinColumn(name = "recipeId")
     private Recipe recipe;
 
-    private Integer stepOrder;
+    private Long stepOrder;
 
     @Column(columnDefinition = "TEXT")
     private String description;
